Create extension to show Travis CI results in a ticket.
This extension is modeled after RT::Site::BPS::Smokingit.  In the
ticked display in appropriate queues, it adds a "Travis-CI Status"
section showing the status of the most recent Travis CI build for
the branch.  As with Smokingit, the project name and branch name
are extracted from the ticket title.

Note that you need to create a Travis CI token and configure it in
this extensions configuration file to be able to fetch data using the
Travis CI API.  Also note that currently, the extension runs
server-side and therefore the page rendering is delayed while the
server fetches the Travis CI results.  For a future version, we probably
want to fetch the Travis CI results using AJAX in the browser.  We will
still need a server-side component to avoid cross-origin permission
problems.